feat: adicionado propriedade quantidade que representa a quantidade do produto por estoque

+++ GestorDeEstoque/Models/ProdutoEstoque.cs
@@ -10,6 +10,6 @@ namespace GestorDeEstoque.Models
         public int EstoqueId { get; set; }
         public Estoque? Estoque { get; set; }
 
-        public int Quantidade { get; set; }
+        public decimal Quantidade { get; set; }
     }
 }
